requirements.txt
faiss-cpu #==1.8.0
numpy==1.24.2
openai #==1.3.3
pandas #==1.3.4  #pandas==1.5.3
python-docx #==1.1.2
PyPDF2 == 3.0.1
python-dotenv
scipy==1.6.0
fastapi
uvicorn == 0.30.6

tiktoken == 0.5.1